Courses work with Floats

+++ src/com/company/CourseInfoPack.java
@@ -3,8 +3,8 @@ package com.company;
 public class CourseInfoPack {
     Integer courseId = null;
     Integer teacherId= null;
-    Integer average = null;
-    CourseInfoPack(Integer id, Integer teacherId, Integer average){
+    Float average = null;
+    CourseInfoPack(Integer id, Integer teacherId, Float average){
         this.courseId = id;
         this.teacherId = teacherId;
         this.average = average;
